+++ utils/__init__.py
@@ -6,6 +6,7 @@ from .cache_manager import CacheManager, cache_manager
 from .async_cache_manager import AsyncCacheManager, async_cache_manager
 from .git_manager import GitManager
 from .websocket_manager import WebSocketManager, websocket_manager, CollaborationSession
+from .monitoring_dashboard import MonitoringDashboard, monitoring_dashboard
 from . import async_helpers
 
 __all__ = [
@@ -19,6 +20,8 @@ __all__ = [
     'WebSocketManager',
     'websocket_manager',
     'CollaborationSession',
+    'MonitoringDashboard',
+    'monitoring_dashboard',
     'async_helpers'
 ]